Add Transaction Feed with multi-token support (#1575)
### Description

I know what you're thinking "1600 lines? That's way too much!" BUT over 1k lines are tests and snapshots and this is already a small subset of the feature. I couldn't find a way of meaningfully reduce the number of changed lines but if you have suggestions on what to move to another PR I'm happy to do it if you think that will help reviewing.

This PR is the first step to add multi-token support to the transaction feed. The relevant changes are mostly in three files:
- `packages/mobile/src/transactions/feed/TransferFeedItem.tsx` (new)
- `packages/mobile/src/transactions/feed/TransactionFeed.tsx` (new)
- `packages/mobile/src/transactions/transferFeedUtils.ts` (added 2 new hooks)

There are a few things missing which I opened new issues for:
- Transaction detail: https://github.com/valora-inc/wallet/issues/1572
- Standby transactions: https://github.com/valora-inc/wallet/issues/1573
- Invite names: https://github.com/valora-inc/wallet/issues/1574

### Other changes

Removed the invites reducer since it was never being written to :( In production right now you see `Unknown` from the get go. I added an issue to fix this.

### Tested

Manually and from unit tests.

### How others should test

Not ready to be tested, we need to implement the listed issues first, but to test we need to check that the transaction feed always show the expected transactions and that we can see the details by tapping on it.

### Related issues

- Fixes #1262

### Backwards compatibility

N/A