Multi token request flow (#1516)
### Description

A new request flow for multiple tokens. Users can request using the flow similar to the new send flow. Amount can be requested in coreTokens - currently cUSD and cEUR. Enabled from the `multiTokenUseSendFlow` feature flag. If this feature flag is disabled the old request flow has will be served instead. 

One strange thing about how requests are implemented is that there is no guarantee that the user will in fact send the requested token, only the requested amount. Tackling this issue if desired should come in a subsequent PR as request creation and retrieval will need to be reworked. Currently Requests are using the legacy send confirmation screen due to the transaction data passed to request.
 
### Other changes
- Changes to test to support the new transactionData model while also supporting the old transactionDataLegacy model.
- Changes to tests to support both models.

### Tested
- Added unit tests
- Updated existing unit tests to use the legacy model where appropriate. 
- Tested locally with multiTokenUseSendFlow flag enabled and disabled.

### How others should test

#### With multiTokenUseSendFlow feature flag enabled on Alfajores
- Complete a request from two accounts on the current version
- Send a request from pervious version and complete the flow on the new build.
- Send a request from this build to a pervious version and complete the flow.

#### With multiTokenUseSendFlow feature flag disabled on Alfajores
- Complete a request from two accounts on the current version
- Send a request from pervious version and complete the flow on the new build.
- Send a request from this build to a pervious version and complete the flow.

### Related issues

- Fixes #1395

### Backwards compatibility

This change is should be backwards compatible.