Add exchange items to feed and feed item detail (#1702)
### Description

Added feed items and details for CELO purchases and sells.

<img width="332" alt="Screen Shot 2021-12-30 at 20 44 33" src="https://user-images.githubusercontent.com/6062888/147794837-935dde4d-8c65-4461-90a5-1c5f95f5de45.png">
<img width="337" alt="Screen Shot 2021-12-30 at 20 44 51" src="https://user-images.githubusercontent.com/6062888/147794840-e2979e46-db35-467a-b0f7-fa4906aa97ce.png">

Updated video after changes:

https://user-images.githubusercontent.com/6062888/148303075-a1fa7307-1cee-49a0-ab89-cec5a1c11590.mov


### Other changes

N/A

### Tested

Manually and with unit tests

### How others should test

With the feed multi token support flag enabled, buy/sell CELO and see that the item shows correctly in the feed. 
Then click it and see that the transaction detail looks correct as well.

### Related issues

- Fixes #1595

### Backwards compatibility

N/A