Fire analytics event when selecting Supercharge in the side menu (#1828)
### Description

I realized when setting up the experiment that if the user opens the Supercharge screen from an option in the side menu then we were not firing the `rewards_screen_opened` event. Added the event for that as well.

### Other changes

N/A

### Tested

Manually

### How others should test

No need for QA to test.

### Related issues

- Fixes #1686

### Backwards compatibility

N/A