Fix broken import for Sentry enabled config variable that was causing some error events to be lost (#1834)
### Description

During a refactor, the SENTRY_ENABLED variable moved to config.ts but in the index.js we were still importing from the previous place. This means that SENTRY_ENABLED is always undefined and we are not sending the root errors to Sentry.

This PR fixes the broken import.

### Other changes

Remove unused sentryEnabled variable.

### Tested

N/A


### How others should test

This doesn't need to be tested by QA team

### Related issues

N/A

### Backwards compatibility

Yes