Allow using any stablecoin for gas fees (#1835)
### Description

We can now use any stablecoin to pay for gas fees :) It ended up taking a bit longer than expected. Here's a related PR in the blockchain-api: https://github.com/valora-inc/blockchain-api/pull/101

One exception to this is reclaiming invites. There's more work that's needed there, right now pending invites in cREAL don't even show up as a in-app notification! I'll fix reclaiming in a future PR. 

### Other changes

I'm trying to phase out the use of `Currency` when we can since it's a static enum in favor of using `coreTokens` from the selector since that allows us to add new stablecoins dynamically without an app update.

### Tested

Manually and with unit tests.


### How others should test

Try and action that signs a transaction:
- Sending
- Inviting
- Signing a tx with Wallet Connect.
- Signing a tx with dappkit (ubeswap)
- Exchanging

All should work normally!

### Related issues

- Fixes #1747

### Backwards compatibility

N/A