Resolve simple-get to version 4.0.1 which contains security patch to fix CI (#1887)
### Description

Our vulnerabilities CI has identified simple-get as a new vulnerability and recommends to upgrade to version 4.0.1. This is the error identified https://github.com/advisories/GHSA-wpg7-2c88-r8xv

Example of the error in CI: https://github.com/valora-inc/wallet/runs/4988071468?check_suite_focus=true

The simple-get releases are not well documented so it is difficult to know which (breaking) changes are in which release. Several of our dependencies are using version 2 and 3 of simple-get (komenci, wallet connect, several celo projects) so we are relying on manual regression tests and the e2e CI to verify that resolving simple-get to v4 works for us.

We need to resolve this to make our CI green again 🟢

### Other changes

N/A

### Tested

N/A


### How others should test

Does not need to be tested explicitly (implicitly covered during regression tests)

### Related issues

N/A

### Backwards compatibility

Yes