Bugfix: Only show token error banner when there was an actual error loading tokens (#1909)
### Description

The error banner was showing too often. Make it only show up when there was an actual error fetching tokens.

### Other changes

N/A

### Tested

Manually forcing errors.

### How others should test

The error banner shows up if there is no cached version of the tokens in the last day and there was an error fetching the tokens.
It's hard to test, but if you turn off wi-fi and open the app for the first time after a while (1d+), the error banner should show up.

Alternatively, I think this can be reproduced by restoring an account and turning the WiFi off just before pressing the 'skip' button in the verification screen

### Related issues

- Fixes #1900

### Backwards compatibility

N/A