Throw error when signing a transaction with a 0 gas price (#1954)
### Description

TXs with a 0 gas price are currently stuck in the txpool.
With this change, we'll be alerted when we're signing such TXs and can get more data about why/when it's happening.

Note: starting with celo-blockchain >= 1.5.0, TXs with a 0 gas price will be evicted instead of sitting in the txpool indefinitely, but still would be good to never send such invalid TXs.

### Tested

Added new unit test and manually trying signing a TX with a 0 gas price.

### How others should test

N/A

### Related issues

- #1417
- https://github.com/celo-org/celo-blockchain/pull/1817 (fix for TXs with 0 gas price not being evicted).

### Backwards compatibility

Yes