Run migration to check verification state (#2013)
### Description

Re-running an old migration to the verification state that we ran to solve a similar issue in the past.

### Other changes

N/A

### Tested

I set the internal state to different situations and ran the migration to see it works as expected.

### How others should test

No real way to test this unfortunately without access to the code.

### Related issues

- Fixes #1906

### Backwards compatibility

N/A