Allow claiming Supercharge rewards (#2041)
### Description

Added to the Supercharge screen the ability to claim rewards manually.

https://user-images.githubusercontent.com/6062888/155240630-e64c1e6c-8b07-486b-957c-69cc3452d590.mov


### Other changes

N/A

### Tested

I created mock distributions and claimed them from the app.

### How others should test

No way to test this without creating a mock distribution, we will create one for the Bugapalloza but I don't know how the QA team can do that :/

### Related issues

- Fixes #2017

### Backwards compatibility

N/A