Merchant Payments (#2136)
### Description

In preparation for celo-connect where we would like the conference participants to pay with valora using the merchant-payments API, here is the PR adding that behaviour (under a feature flag).

### Other changes

Nothing normally

### Tested

https://user-images.githubusercontent.com/8698308/157790697-c4d318a1-d2a1-4f2e-9e03-08e0f9d4879b.mov

### How others should test

We have a test shop here: https://sandbox-celo-demo-store.demo.firstdag.com/

### Backwards compatibility

This new feature is behind a feature flag, should be 100% backwards compatible.