Open dapps in a webview (#2144)
### Description

In this PR:
- add remote config flag for `dappsWebViewEnabled`
- if `dappsWebViewEnabled`, open dapp in a WebView instead of an external browser
- Cookies are preserved by default 
- adjust the screen styles for WebView to match [designs](https://www.figma.com/file/zdK8XvxNXvTjxnM5V33pHw/Dapps-Page?node-id=1999%3A45353), add basic bottom nav bar and refresh control

![Screenshot_1647349346](https://user-images.githubusercontent.com/20150449/158383462-2bdd881a-c7a3-46f1-a653-fac0c5f28451.png)


**NOT included** in this PR:
- connecting dapps to app (i.e. clicking "connect wallet" on dapps inside the webview - this requires some more work and will be part of https://github.com/valora-inc/wallet/issues/2101)

### Other changes

Deleted an unused file (InAppBrowser)

### Tested

Manually and some unit test for webview screen and opening dapps via webview or deeplink


### How others should test

Open dapps from dapp explorer/recently used dapps, verify that the dapp is opened in-app.
should be able to navigate back and forth within the dapp using the bottom bar nav controls (e.g. click on a button inside the dapp which leads you to a different part of the dapp, then pressing the back button on the bottom nav bar should return you to the previous dapp page)


### Related issues

- Fixes #2100 

### Backwards compatibility

Yes