fix: bottom sheet not displaying second time after dismiss (#2254)
### Description

The bottom sheet was not displaying a second time after being dismissed as described in #2245. This aims to fix that issue on wherever the `BottomSheet.tsx` is used such as the Dapps page and token bottom sheet.

### Other changes

N/A

### Tested

Tested locally on Android 10 & Android 5 in dev and release mode.

### How others should test

Follow the reproduction steps for #2245 and ensure that the issue does not reproduce.
Follow the reproduction steps for #1973 and ensure that the issue does not reproduce.
Follow the testing steps for #2230.

### Related issues

- Fixes #2245

### Backwards compatibility

Yes