E2E Timeout Fixes (#2347)
### Description

A few e2e tests such as `Ramps.spec.js` are failing due to slow load times this pr adds implicit waits for those elements to render. It also wraps the exchange scroll in a try catch to allow tests to pass even if fetching transactions is slow from blockscout.

### Other changes

N/A

### Tested

Tested in CI.

### How others should test

This does not need to be tested by QA. Reviewers should view the e2e test runs on this branch.

### Related issues

N/A

### Backwards compatibility

Yes