Update auto-populated invite text for invite rewards users (#2400)
### Description

Details in #2291 

This PR adds a condition where if invite rewards are active for a user, after inviting a friend the auto populated text message content should mention the reward. For all other users, there should be no mention of this reward (i.e. the status quo)

### Other changes

N/A

### Tested

Manually + with unit tests


### How others should test

For users who are located in the US + when the invite rewards program is active (only true on Alfajores right now), the text they should see is "Hey! I’ve been using and enjoying Valora as a crypto wallet. If you download the app and connect your phone number, we’ll both get 5 cUSD. Use this link: [some invite link that links to the download page of the valora website]"

For all other users/when the invite rewards program is not active, there should be no mention of the 5 cUSD reward in the auto populated text message.

### Related issues

- Fixes #2291 

### Backwards compatibility

_Brief explanation of why these changes are/are not backwards compatible._