fix: incorrect token shown when opening a transfer push notification (#2480)
### Description

This fixes the token shown when opening a transfer push notification, when the token sent is not cUSD.

Related change: https://github.com/valora-inc/notification-service/pull/21

### Tested

**before/after**

Sending cEUR to an account and opening the push notification:
<img src="https://user-images.githubusercontent.com/57791/168777115-d8b5d71b-f745-4cac-8592-c87ee096c394.png" width="50%" /><img src="https://user-images.githubusercontent.com/57791/168777156-9ee1dbe3-a64b-4558-93cd-42b8f83d6f81.png" width="50%" />

Note: this was on Alfajores were the EUR/USD rate is a bit funky.

### How others should test

1. Send a token which is not cUSD (for instance cEUR or [TT](https://test-token-faucet.web.app/)) to another account
2. On the receiving account device, tap the push notification
3. The opened screen should show the actual token sent, and not cUSD.

### Related issues

- Fixes #1968

### Backwards compatibility

Yes (requires https://github.com/valora-inc/notification-service/pull/21 to be deployed).