chore: migrate existing geth + dek accounts to the Keychain based wallet (#2539)
### Description

Builds on top of #2519 to add backward compatibility for existing user accounts.

This transparently migrates the existing geth + dek accounts into the Keychain based wallet.

The geth account is re-imported from the mnemonic the next time the user enters their PIN to perform an action that requires it (signing, updating their PIN, etc).

The dek is re-imported on the first launch if a user account is already set. It doesn't require the PIN (by design).

### Tested

- Updated unit tests for the KeychainWallet
- Manually upgrading the app from the the geth based wallet to the new Keychain based wallet on both iOS and Android.

### How others should test

1. Have an existing app <= v1.33 with a user account set
2. Ensure TXs can be sent and and the seed phrase can be seen.
3. Update the app with version >= 1.34 (containing this commit)
4. Make sure all things related to signing work as expected
    - sending transaction
    - swaping
    - signing with dapps
    - phone number verification
    - updating the PIN

### Related issues

- Part of #2392 

### Backwards compatibility

Yes